emedicare plus/backend/requirements.txt
Flask>=2.0
flask_cors>=3.0
mysql-connector-python>=8.0
werkzeug>=2.0
